spv: Add non-wallet-backed mempool to Syncer
This adds a mempool on the SPV Syncer so that it can actually relay
transactions that are not relevant to the wallet.

Transactions that are relevant (i.e. stored in the wallet's DB) don't
need to be stored in the mempool since they are already accessible when
a GetData request comes in from a peer.